requirements.txt
backcall==0.1.0
backports.csv==1.0.7
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.7
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-extensions==2.1.6
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.2.0
django-redis==4.10.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.3
idna==2.7
ipython==7.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.1
Markdown==3.0.1
MarkupSafe==1.0
mysqlclient==1.4.2
odfpy==1.4.0
openpyxl==2.6.2
parso==0.4.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
Pygments==2.3.1
PyJWT==1.6.4
PyMySQL==0.9.3
pytz==2018.6
PyYAML==5.1
redis==3.0.1
requests==2.21.0
six==1.11.0
tablib==0.13.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.2
wcwidth==0.1.7
xlrd==1.2.0
xlwt==1.3.0
